Support to delete multiple keyword hits results

+++ DirectoryTree/src/org/sleuthkit/autopsy/directorytree/ResultDeleteAction.java
@@ -21,10 +21,11 @@ package org.sleuthkit.autopsy.directorytree;
 import java.awt.event.ActionEvent;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.swing.AbstractAction;
-import org.openide.util.Exceptions;
+import javax.swing.JOptionPane;
 import org.sleuthkit.autopsy.casemodule.Case;
 import org.sleuthkit.datamodel.BlackboardArtifact;
 import org.sleuthkit.datamodel.SleuthkitCase;
@@ -33,36 +34,43 @@ import org.sleuthkit.datamodel.SleuthkitCase;
  * Action that deletes blackboard artifacts requested and reloads the view
  */
 public class ResultDeleteAction extends AbstractAction {
-    
-    enum ActionType {
-        SINGLE_ARTIFACT  ///< deletes individual artifacts and assoc. attributes
+
+    private enum ActionType {
+
+        SINGLE_ARTIFACT ///< deletes individual artifacts and assoc. attributes
+        ,
+        MULT_ARTIFACTS ///< deletes multiple artifacts and assoc. attributes
         ,
         TYPE_ARTIFACTS ///< deletes all artifacts by type and assoc. attributes
     }
-    
     private BlackboardArtifact art;
     private BlackboardArtifact.ARTIFACT_TYPE artType;
+    private List<BlackboardArtifact> arts;
     private ActionType actionType;
-    
     private static final Logger logger = Logger.getLogger(ResultDeleteAction.class.getName());
 
     ResultDeleteAction(String title) {
         super(title);
     }
-    
+
     ResultDeleteAction(String title, BlackboardArtifact art) {
         this(title);
         this.art = art;
         this.actionType = ActionType.SINGLE_ARTIFACT;
     }
-    
+
+    ResultDeleteAction(String title, List<BlackboardArtifact> arts) {
+        this(title);
+        this.arts = arts;
+        this.actionType = ActionType.MULT_ARTIFACTS;
+    }
+
     ResultDeleteAction(String title, BlackboardArtifact.ARTIFACT_TYPE artType) {
         this(title);
         this.artType = artType;
         this.actionType = ActionType.TYPE_ARTIFACTS;
     }
-    
-    
+
     @Override
     public void actionPerformed(ActionEvent e) {
         if (actionType == ActionType.SINGLE_ARTIFACT) {
@@ -70,14 +78,26 @@ public class ResultDeleteAction extends AbstractAction {
             DirectoryTreeTopComponent viewer = DirectoryTreeTopComponent.findInstance();
             viewer.refreshTree(BlackboardArtifact.ARTIFACT_TYPE.fromID(art.getArtifactTypeID()));
         }
-        else if (this.actionType == ActionType.TYPE_ARTIFACTS) {
-            deleteArtifacts(artType);
+        if (actionType == ActionType.MULT_ARTIFACTS) {
+            for (BlackboardArtifact art : arts) {
+                deleteArtifact(art);
+            }
             DirectoryTreeTopComponent viewer = DirectoryTreeTopComponent.findInstance();
-            viewer.refreshTree(artType);
+            viewer.refreshTree(BlackboardArtifact.ARTIFACT_TYPE.fromID(art.getArtifactTypeID()));
+        } else if (this.actionType == ActionType.TYPE_ARTIFACTS) {
+            if (JOptionPane.showConfirmDialog(null,
+                    "Are you sure you want to delete all " + artType.getDisplayName() + " results?",
+                   artType.getDisplayName() + " Results Deletion", JOptionPane.YES_NO_OPTION,
+                    JOptionPane.QUESTION_MESSAGE) == JOptionPane.YES_OPTION) {
+                deleteArtifacts(artType);
+                DirectoryTreeTopComponent viewer = DirectoryTreeTopComponent.findInstance();
+                viewer.refreshTree(artType);
+            }
+        } else {
+            throw new IllegalArgumentException("Invalid action type: " + this.actionType);
         }
-        else throw new IllegalArgumentException("Invalid action type: " + this.actionType);
     }
-    
+
     //TODO should be moved to SleuthkitCase and BlackboardArtifact API
     private static void deleteArtifact(BlackboardArtifact art) {
         final SleuthkitCase skCase = Case.getCurrentCase().getSleuthkitCase();
@@ -88,32 +108,32 @@ public class ResultDeleteAction extends AbstractAction {
 
             rs = skCase.runQuery("DELETE from blackboard_artifacts where artifact_id = " + Long.toString(artId));
             skCase.closeRunQuery(rs);
-            
+
         } catch (SQLException ex) {
             logger.log(Level.WARNING, "Could not delete artifact by id: " + artId, ex);
         }
-        
+
     }
-    
+
     //TODO should be moved to SleuthkitCase
     private static void deleteArtifacts(BlackboardArtifact.ARTIFACT_TYPE artType) {
         // SELECT * from blackboard_attributes INNER JOIN blackboard_artifacts ON blackboard_artifacts.artifact_id = blackboard_attributes.artifact_ID AND blackboard_artifacts.artifact_type_id = 9;
         final SleuthkitCase skCase = Case.getCurrentCase().getSleuthkitCase();
         try {
-            ResultSet rs = skCase.runQuery("DELETE FROM blackboard_attributes WHERE artifact_id in "
+            ResultSet rs = skCase.runQuery("DELETE FROM blackboard_attributes WHERE artifact_id IN "
                     + "(SELECT blackboard_artifacts.artifact_id FROM blackboard_artifacts "
                     + "INNER JOIN blackboard_attributes ON (blackboard_attributes.artifact_id = blackboard_artifacts.artifact_id) "
                     + "WHERE blackboard_artifacts.artifact_type_id = "
+                    + Integer.toString(artType.getTypeID())
+                    + ")");
+            skCase.closeRunQuery(rs);
+
+            rs = skCase.runQuery("DELETE from blackboard_artifacts where artifact_type_id = "
                     + Integer.toString(artType.getTypeID()));
             skCase.closeRunQuery(rs);
 
-            rs = skCase.runQuery("DELETE from blackboard_artifacts where artifact_type_id = " 
-                    + Integer.toString(artType.getTypeID()));
-            skCase.closeRunQuery(rs);
-            
         } catch (SQLException ex) {
             logger.log(Level.WARNING, "Could not delete artifacts by type id: " + artType.getTypeID(), ex);
         }
     }
-    
 }
